api-server/requirements.txt
django
psycopg2
mysqlclient
channels
djangorestframework
markdown
django-filter
channels_redis
redis
daphne
Pillow
djangorestframework_simplejwt